Fix for GRAILS-2461: the GrailsWebRequestFilter is no longer misused.
In addition to this fix, several aspects of the URL mapping and request processing have been modified to improve consistency and fix minor issues. The "controller" and "action" parameters are now always available in the "params" object, as are any explicit parameters provided in the URL mappings. These are also available to controllers that response codes map to.

git-svn-id: https://svn.codehaus.org/grails/trunk@7012 1cfb16fd-6d17-0410-8ff1-b7e8e1e2867d