add support for customizing the GrailsWebApplicationContext
Currently, there is no way to customize creating and configuration of the GrailsWebApplicationContext. This patch allows to specifiy a context-param in web.xml for a different implementation of GrailsRuntimeConfigurator. 

In order to create a different GrailsWebApplicationContext, a configurator derived from GrailsRuntimeConfigurator may create a spring configuration derived from WebRuntimeSpringConfiguration, which in turns creates a different implementation of GrailsWebApplicationContext.

Use case: OSGi support (using Spring DM) requires that the ApplicationContext implements org.springframework.osgi.context.ConfigurableOsgiBundleApplicationContext, so a class which is derived from GrailsWebApplicationContext and implements ConfigurableOsgiBundleApplicationContext is necessary. The OSGi plugin will supply such a class, and this patch is required so it may be created and used by Grails.