Revert "GRAILS-6790 - register the auto proxy creator in the “root” application context so that any configuration made by plugins/apps apply to our proxy creator."
Installing the APC in this way makes it not be used when generating transactional proxies, which means that Groovy classes don't automatically get proxied by class. Unfortunately, installing it in a way that does allow it to be used when generating transactional proxies runs up against the Spring bug SPR-7808 (which prevents usage of custom APC with 'aop' or 'tx' namespaces).

This exercise has revealed that any user or plugin code that looks up the APC via it's well defined name during application bootstrap will not receive our custom instance. When Spring is doing this internally it will create an instance based on the default implementation. Once the Spring bug is fixed, the solution for us will be to replace line 72 of this patch we are reverting to:

getSpringConfig().getUnrefreshedApplicationContext().beanFactory.registerBeanDefinition(AopConfigUtils.AUTO_PROXY_CREATOR_BEAN_NAME, proxyCreatorDefinition)