Configure the default chain resolver to treat all versions ending in SNAPSHOT as changing, thereby delegating resolution to the actual resolvers.
This is necessary because the chain resolver will check the cache before trying the real resolvers, so if the dependency is in the cache the resolvers will never get a chance to look for updated versions.

This does mean however that if a particular resolver has a changing pattern that is more restrictive than the chain resolver's, it is going to be ineffectual.