fix for GRAILS-7099 "Accept header should take precedence over Content-Type in content negotiation"
Note that there is a new property on the response called 'format' which uses either the file extension, a format request parameter or the ACCEPT header to decide on the response format.

The withFormat method has been changed to use this response.format property. This is a breaking change, but an valid one. The existing request.format property has been changed to only evaluate the CONTENT_TYPE header.

There are also new methods added to the request and response objects.

request.withFormat { } uses request.format and response.withFormat { } is the same as the withFormat method for controllers