Improve the handling of the groovy.jar system property
At build time the groovy.jar system property may be set to point to a groovy jar file which should be used for this build.  The joint build takes advantage of this.  When the property points to a jar file which does not exist, errors occur and it isn't always clear what the cause of those errors is.  This change fails the build with a specific error message that describes the problem.