GRAILS-8343 - NoSuchMethodError at command object validation
When a controller action accepts a parameter that is of a type other than String, a primitive or a type wrapper Grails will treat the type as a command object class.  Command object classes are made validateable at compile time.  A problem occurs if the class that is being used as a command object class is not being compiled as part of the project.  For those cases, we can't compile in the validation support.  This commit identifies situations like that and throws a compilation error instead of letting that turn into a runtime NoSuchMethodError when we try to call .validate().  If the class is not part of the project but is marked with @Validateable, this problem does not occur since the .validate() method will already be there.