New command object handling for Grails 2.1
As of Grails 2.1 command objects will no longer automatically be made validateable.  If a command object class is not marked with @Validateable then the controller action will still accept the command object and will still do data binding but .validate() will not be invoked.  A relevant compiler warning will be issued.