GRAILS-8759 - fix implicit @TestFor
The code that looks at a test class name like AuthorControllerTests to figure out if @TestFor(AuthorController) should be implicitly added was checking to see if the test class name ends with the artifact indicator.  It was doing something like if('AuthorControllerTests'.endsWith('Controller')) instead of if('AuthorController'.endsWith('Controller')).