GRAILS-8853 - fix "help create-app"
The "grails help create-app" command was showing a usage statement for the integrate-with command.  The problem is that the help command is evaluating the CreateApp_ script to try and get the USAGE variable.  CreateApp_ didn't define a USAGE string, which normally does not cause a problem except that the CreateApp_ script indirectly evaluates the IntegrateWith script, which does define a USAGE string so that value was being used by the help command.  This commit addresses the problem for both create-app and create-plugin.