Fix GRAILS-8950: overwrite existing CSS & js files when generating docs.
Although DocPublisher would overwrite CSS, js, and image files that the user provided, it wouldn't overwrite the existing files with the ones that come from grails-doc-files.jar. When a new JAR with new files came along, the user still ended up with the old files in the generated docs unless he did a clean. So, all copies are now overwrite: true.