Also cache negative uri cache lookups.
Serving views or resources which do not have a controller is pretty slow due to uri lookups on every request.
This fixes it by also caching uri looks which do not have a controller.

To avoid DOS attacks, the cache is also limited in size. Since there's no
caching framework available currently, the cache has been modified to a simple
fifo cache based on LinkedBlockingQueue.