Also cache negative uri cache lookups.
Serving views or resources which do not have a controller is pretty slow due to uri lookups on every request.
This fixes it by also caching uri looks which do not have a controller.

To avoid DOS attacks, the cache is also limited in size, and the size can
be controlled via the grails.urlmapping.cache.maxsize config property.