Modified forward dynamic method to use UrlConverter for controller/action parameters.
This fix solves a 404 error when using the hyphenated URL converter in
conduction with a forward request.  The redirect method already uses
the converter transitively through the LinkGenerator (which the forward
method does NOT use).  This patch brings the forward method in line
with the logic applied by redirect and fixes the 404 issue when the
forward URI is matched against the URL mappings.