Prevent url encoding for controller and action
Currently, the controller name and action name are url-encoded when creating an HTML link.
But I want to be able to configure my controller to have a featureUri with a slash in it - e.g. "admin/myController". This should be possible using the UrlMappings framework, but currently fails because the slash is encoded to "%2F" by the LinkGenerator. So the URI becomes "admin%2FmyController".
This patch prevents URL-encoding the controller name and action, when creating a URL.