don't retrieve the value from the binding Map until/unless it is needed
Using Map.each { key, value -> … was retrieving every value from the Map, even those that aren't needed.  For normal Maps this isn't an issue but to optimize the use of BeanMap (and potentially other lazy Map types) this change defers retrieving the value from the Map until/unless it is needed.