GRAILS-10220 - handle XML or JSON body parsing errors
If the parsing error occurs during the automatic data binding that is imposed on command objects, the response status code is set to 400 and the controller action is never actually invoked.  If the parsing error occurs during binding initiated by application code, a binding error is added to the object that data binding is being imposed upon.