GRAILS-10091 - Improve service bean aliasing
The framework is only supposed to create aliased bean names for Service artifacts if the alias name will not conflict with another bean. This worked if the conflicting bean is a Service artifact. It did not work if the conflict was created by a bean that is not a Service artifact but is a bean explicitly added to the context.

This change addresses that problem by moving the alias creation to later in the process after the context is initialized so the search for conflicts can include all beans, not just Service artifact beans.