GRAILS-5779 - restore the original bean builder
The bean builder work isn't complete but all of the code is working for
me locally. spring-lang-groovy hasn't been published to any public repo yet so I am rolling this back until that is available.  I want to get the rest of the Spring 4 code pushed out and running in the mean time so we can work through issues.