GRAILS-10448 - fix singleton property access in the bean builder
The bean builder supports a syntax that allows the "singleton" property of the bean configuration to be assigned a boolean. Prior to Spring 4 this worked because AbstractBeanDefinition.setSingleton() existed. That method has been removed in Spring 4 so we need special handling to support the same syntax in the bean builder.

The version of the bean builder that we are working on for inclusion in Spring 4 also contains this commit will be irrelevant once we get to that point.