Test argument to create-multi-project-build for validity.
Fixes GRAILS-10494. If the argument to `create-multi-project-build` didn't
consist of 3 parts separated by colons, the command just threw an exception
when trying to assign the result of `split()` to a tuple. I have added an
explicit check for the 3 parts and simply print a message now and exit if
the number of parts is incorrect.