Refactor Maven POM generation for consistency.
The number of POMs has now been cut from 5 or 6 to just 2. This makes it much
easier to maintain them and ensure that all generated POMs are up to date. As
part of this refactor, I have also switched from Ant file filtering to Groovy's
SimpleTemplateEngine, which allows for more flexibility in the templates.

There is still a problem in that the `create-multi-project-build` command can't
produce the same POMs in child projects as the `create-pom` command. This is
because you can not get the dependencies of an arbitrary project, only the one
that a command is run in. The best way to fix this is to allow for dependency
resolution on any Grails project, no matter where it is in the file system. And
of course you have to be able to resolve the dependencies for multiple projects
in one JVM instance.