Clarify why 'input disabled' error occurs for user.
When Grails thinks that a new version of a plugin needs installing, it asks the
user whether to install that new version. This is fine when using the `grails`
command, but if the project is being built with Maven, it causes an exception.
Previously, this exception gave no indication what caused it.

This change incorporates the 'prompt' that caused the exception in the error
message, thus allowing users to determine how to fix it (perhaps by fixing
plugin versions in POMs or cleaning the target directory).