Put wrapper-support jar in grails-resources jar.
Fixes GRAILS-10653. The `_GrailsWrapper.groovy` script now just unpacks the
grails-wrapper-support jar from the classpath, and that jar now also includes
the springloaded jar. This matches with the way that new project structures
are created. In fact, these files should probably be part of the
grails-shared-files jar instead, but that would have required more changes.