GRAILS-8426 - improve allowedMethods handling
This change moves some of the allowedMethods handling from runtime to compile time.  Controller actions which are specified in allowedMethods will have code added to the top of the method which does the allowedMethods check.  The old runtime check code is still there to deal with controllers which were compiled with versions of Grails prior to 2.4 but that code is skipped for controllers that are compiled with 2.4 and later.  The 2.4 compiler is adding the @AllowedMethodsHandledAtCompileTime annotation to controller classes at compile time so these classes can be identified at runtime.