Bind the server listening socket to the address specified in serverHost
Some server configurations require the HTTP server to listen on a specific
address or interface. Until now, there was no way to specify an address to
bind the listening socket to - the serverHost variable was cosmetic.
Grails now uses the serverHost variable, if it is specified, as the bind
address of the HTTP/HTTPS listening sockets.