GRAILS-11462 - fix REST binding problem
Updates are being done with the request parameters, even if the request has a body. The fix is to change that so when the update is done, if the request has a body then the body is used and if it does not, the request a parameters are used.

This is related to https://jira.grails.org/browse/GRAILS-11354.