GRAILS-11601 -simplify some constraint related logic
This simplifies some logic around the code which decides which properties should be forced to be constrained at compile time.  The prior logic was discovering all setter methods and then only forcing the property to be constrained if a corresponding getter method was found.  This shortcuts that and just looks for getter methods.  The setter shouldn't be relevant when deciding if a property is constrainable.

This change is related to both GRAILS-11601 and GRAILS-9686.