GRAILS-11661 - fix unit test metaclass cleanup issue
This is related to the change made at https://github.com/grails/grails-core/commit/1885351d2be8f74076f2c83edcbc048c5bcfd9d7.  The change in this commit changes the runtime to fire a resetGrailsApplication event in addition to the after event.  I am not yet sure if this is the correct thing to do.  Pending investigation.