GRAILS-11686 - improve request data binding
When a request is used as a data binding source and the content type indicates that the request has a body normally the body is used during data binding instead of request parameters.  This change changes things such that if the content type indicates that the request has a body but the body is empty then request parameters will be used for data binding.

This is related to GRAILS-11263 as well as GRAILS-11686.