fix failing test
In Grails 2 the "properties" property was made writeable for all classes which were used as command objects.  I think this was accidental/incidental and not really intentional.  The previous commit caused tests to fail which pointed this out.  If we really want the "properties" property to be writeable we can move the setProperties method out of grails.artefact.Domain and into some common trait which is applied to domain classes and command objects alike.