temporarily introduce TempControllerServletApi
This is a temporary step to start isolating parts of the controller api which have direct references to the servlet api.  This class is not intended to persist in its current state but this step will be helpful in progressing toward an arrangement that better supports Grails controllers in environments that are not a servlet container.