Use GrailsASTUtils helper method
We need to reorganize how some of this code works.  Right now there are 2 entry points into ArtefactTypeAstTransformation.  When triggered directly by the groovy compiler the visit method is invoked.  When triggered by GlobalGrailsClassInjectorTransformation the static performInjection method is invoked.  The visit method already has a check to see if the transformation has been applied and will quickly return if it has.  A problem occurs if the GlobalGrailsClassInjectorTransformation is run after ArtefactTypeAstTransformation is run because the isApplied check doesn't happen when GlobalGrailsClassInjectorTransformation invokes performInjection.  This change adds the isApplied check to performInjection and uses a different key (the injector class instead of the transformation class).