improve handling of invalid constrained property states
Methods like isEmail(), isUrl() etc. in ConstrainedProperty were coded to throw an exception if the constrained property was not a String type.  We should throw an exception if one of those constraints is applied to a non String type but it should be considered valid to ask a ConstrainedProperty if it is one of those, even if the property is not a String.

This is tangentially related to GRAILS-12010.