Fix problem with formatDate and StreamCharBuffer
It isn’t clear under what circumstances the format might be a
StreamCharBuffer and that might or might not be a separate problem, but
this makes the problem go away.  There really isn't anything peculiar about this particular attribute so whatever scenario leads to this problem for the format attribute could probably happen for attributes in general.

Fixes #9159