Issue #685 - Second format in DateConversionHelper should also use HH
Lenient Calendars will correctly parse hours 13-23 using either 'hh' or
'HH' and lenient is the default.  Since the formats are used to parse date
strings and not format dates, this change should not be a breaking change
or have any noticeable impact.  However, for consistency it make sense to
change to match the other hour formats.