Change how command objects are initialized
Prior to this change if a controller was written like this:

    class SomeController {
        def someAction(SomeDomain sd) {
            // application code here
        }
    }

The compiler would turn that into something like this:

    class SomeController {
        def someAction() {
            // oversimplified here for brevity...
            SomeDomain sd = SomeDomain.get(params.id)
            sd.validate()
            someAction(sd)
        }
        def someAction(SomeDomain sd) {
            // application code here
        }
    }

After this change the compiler will generate something like this:

    class SomeController {
        def someAction() {
            someAction(null)
        }
        def someAction(SomeDomain sd) {
            // oversimplified here for brevity...
            if(sd == null) {
                sd = SomeDomain.get(params.id)
            }
            sd.validate()

            // application code here
        }
    }

The initialization code has been moved into the original method and the null check is there to accomodate unit tests which explicitly pass a parameter in to the action method.

This is related to grails/grails-data-mapping#580.