Improve compiler warning handling
This is related to the recent fix for grails/grails-data-mapping#580.
There isn’t really a runtime behavior problem here.  The mechanism that
warns about invalid command object types just isn’t aware of the
transaction handling logic so it is flagging a concern about an attempt
to use `TransactionStatus` as a command object type.

Fixes #9770