improve domain class detection
One problematic scenario is when a domain class is used as a command
object, the SourceUnit that is being used to figure out if this is a
domain class or not is the source unit of the controller, not the
source unit of the domain class so the previous logic wasn’t working.

It looks like something changed that affects the order that
transformations are being run because in the scenario described above,
the domain class does not yet have the Entity annotation added to it so
it isn’t being detected as a domain class in the same way it was with
previous versions of Grails, at least in some circumstances.

Relevant functional test:
https://github.com/grails/grails3-functional-tests/commit/a115c7928b4064
e6e03eb6a331f4be5d4dd800d6

fixes #10079