Simplify configuration of GSP. Fixes #9936
GroovyPagesGrailsPlugin was reading all plugin gsp/views.properties. This is unnecessary as these files were already read in the BinaryGrailsPlugin class and looked up via GroovyPageLocator from the binary plugins. This changes removes that logic and also adds a change to use classpath:gsp/views.properties which fixes the resolution of the default views.properties for unexploded deployment and for Jetty. See #9936