Optimize message bundling scanning
Searching for message bundles *.properties was scanning the whole classpath which took 2-3 seconds of startup time. This change optimises the scanning to only find bundles relative to the application class. A "searchClasspath" property has been added to the "messageSource" bean to restore the previous behaviour but in 95% of cases this will not be necessary.