Don't initialize the transaction manager until a bean is found that implements TransactionManagerAware. Fixes https://github.com/grails/gorm-hibernate5/issues/7
The TransactionManagerAwarePostProcessor was causing early initialization of all of GORM which meant GORM classes could not using @Autowired to process dependencies. This fixes that by only initiasing the transaction manager if a bean is found that implements TransactionManagerAware. This will be late enough in the process that by then the autowire bean post processor will have been registered.