Deprecate support for Reactor 2 promises. Fixes #10247
Since Reactor 2 is largely abandoned and Reactor 3 is a complete re-write, it has been decided that Reactor will be completely pulled from the codebase in the future and the Events framework re-written on-top of either RxJava or the regular Spring event publisher. Given this stance it makes sense to deprecate all related Reactor classes now so that users can consider this migration for the future.