Improve property detection
GrailsClassUtils.getPropertyForGetter/Setter would previously return property names for some method names that were not valid
getters or setters, although the documentation says they return null in those cases. This change improves the detection
and makes sure only valid getters/setters result in a property name.

This commit includes updates in the documentation of related methods to more clearly indicate what is accepted and what not,
and what assumptions are made.

This also includes a bunch of new test cases.