Re-use GrailsNameUtils property logic instead of separate logic
Re-use GrailsNameUtils.getPropertyForGetter from ClassPropertyFetcher instead of having
separate (and wrong) logic.

This commit has a small behavioral change in that the old code would register both
"Property" and "property" for a static property with a getter "getProperty" in
ClassPropertyFetcher. With improved property name derivation that should not be
necessary anymore.