Ignore migrations folder when getting project package names
If a project uses the grails database-migration plugin, by default, the
changelog and changesets are written to the grails-app/migrations/
folder. These files can be either xml or groovy scripts. If they're
groovy scripts, they'll be picked up when scanning for packages. On app
startup the following error will be logged by ClassPathScanner because
the scripts are interpreted to be in the default package:

  "The application defines a Groovy source using the default package.
  Please move all Groovy sources into a package."

Since these migrations aren't application classes, they can be ignored
from scanning.