Upgrade React profile, added React-Webpack profile
React-Webpack was formerly the "1.x" branch of the React profile. This change is to allow users to specify either version of the profile by name when creating a project (rather than having to include the full artifact group/name/version). In addition both profiles have been upgraded to React 16.x.