Replace @Bean with @Named @Singleton in MimeTypeConfiguration
This bean will injected from MicronautApplicationContext, as it is the parent for Spring ApplicationContext. So, we can safely add the bean only in the parent context. In this change we are replacing `org.springframework.context.annotation.Bean` with `javax.inject.Named` and `javax.inject.Singleton`. Also, we do not need `org.springframework.context.annotation.Primary`.

It also include deprecating `MimeTypesHolder` as the only purpose was to register `MimeType[]` singleton which was not possible earlier with `@Bean`.

Add license information to the classes.