Fix @GrailsCompileStatic on domain objects with inner classes (#13013)
Make sure that for every class a new scope is added to and removed from the stack.

Previously a new scope was added for nearly every class but not removed.
This had the effect that the wrong (empty) scope (from the inner class)
was used for the entity class. This was then missing any of its `mapping`,
`constraints` and `namedQueries` closures.