fix(spring6): Config loads correctly to Spring profiles
Spring has introduced a new `spring.config.activate.on-profile` config
property to signal if a config document should be included or not.

This commit refactors the existing `shouldSkipBlock` method to take
that into consideration. However, the usage of `shouldSkipBlock` was
removed in commit 03bef04. I'm not sure if that was an oversight, but
I have added the usage of this method back (now named `springProfileExclude`).