feat!: Update to Spring 6, Boot 3 (#13545)
* feat(spring6)!: Move from javax.servlet to jakarta.servlet
* chore: Update license headers
* fix(spring6): Simplify BeanCreationProfilingPostProcessor
- Remove overridden methods that has the same behavior as the default implementations.
- `postProcessPropertyValues()` is removed in Spring 6 and replaced with `postProcessProperties()` but this is now removed.
- Add @Nullable and @NonNull method parameter annotations to remove warning in Intellij.
* fix(spring6): Remove deprecated OptimizedAutowireCapableBeanFactory

This class is deprecated and uses API removed in Spring 6 -
`AbstractBeanFactory.getAccessControlContext()`

It also uses deprecated API in Java 17 - `java.lang.AccessController`
and `java.lang.SecurityManager`

* fix(spring6): Remove grails-web-fileupload project

The Spring CommonsMultipartResolver is removed in Spring 6.

The grails-web-fileupload project seems to have been created to address
a specific problem with the Safari browser. Hopefully this is resolved now
and this fix no longer necessary.

* fix(spring6): Deprecate MockHttpServletResponse

The overridden methods in this class has been removed in Spring 6.
This class does not serve a purpose any longer and is therefore deprecated.

* fix(spring6): Replace removed NestedIOException

NestedIOException has been removed in Spring 6.
IOException should be used directly instead.

* fix(spring6): Add missing classes for annotation reading

The class hierarchy for AnnotationMetadataReadingVisitor has been
removed in Spring 6 with no public replacement. This commit adds
them to the grails-core project instead as a workaround until
another solution is found.

* fix(spring6): Config loads correctly to Spring profiles

Spring has introduced a new `spring.config.activate.on-profile` config
property to signal if a config document should be included or not.

This commit refactors the existing `shouldSkipBlock` method to take
that into consideration. However, the usage of `shouldSkipBlock` was
removed in commit 03bef04. I'm not sure if that was an oversight, but
I have added the usage of this method back (now named `springProfileExclude`).

* test: Set `@PendingFeature` on tests that uses `grails-plugin-converters`

`grails-plugin-converters` has not yet been converted to jakarta.servlet.
This commit makes the test suite pass until that is fixed.

`GrailsMockHttpServletRequestTests` had to be converted from JUnit to Spock to be able to use `@PendingFeature`.

* test: Disable tests using `grails-gsp`

The `grails-gsp` module has not yet been converted to jakarta.servlet,
causing test failures. To ensure the test suite passes until this
conversion is complete, this commit disables the affected tests using
`@Ignore` (Spock) and `@Disabled` (JUnit5).

As the problem with the tests occurs in the initialization phase,
it was regrettably not possible to use `@PendingFeature`.

* test: Change Spring profile property in test config

Use the new `spring.config.activate.on-profile` config property
instead of the old `spring.profiles`.

* fix(deps): Update Spring to 6.0.22

* fix(deps): Update Tomcat to 10.1.25

* fix(deps): Update Spring Boot to 3.1.12

* feat!: Update to and compatible with Spring 6.1, Boot 3.2

* test(deps): Remove unused test dependencies

* feat!(deps): Update to and compatible with Micronaut 4.5

As Micronaut 4 does no longer include `jackson-databind` as a transitive
dependency, it is added as a direct dependency where ObjectMapper is used.

* chore(deps): Remove unused test dependencies

* docs: Add javadoc explaining that classes are ported from Spring