ci: Remove check for release-drafter.yml
The logic looking for a `release-drafter.yml` file and actions handling if that is not found seems to trigger even though the file is there.

As we have a `release-drafter.yml` file in this project this logic seems unnecessary.

This commit removes it, to see if this fixes the problem with the workflow.