compile grails-shell, grails-gradle-model and grails-bootstrap w/ Gradle Groovy version (#13653)
* compile grails-shell, grails-bootstrap and grails-gradle-model with the Groovy version provided by Gradle

* fix: get `groovydoc` task working again

Reverts some changes that updated to Java 17 syntax `instanceof`. At the moment this breaks groovydoc generation with the message:
Use of patterns with instanceof is not supported. Pay attention that this feature is supported starting from 'JAVA_14' language level. If you need that feature the language level must be configured in the configuration before parsing the source files.

As these changes are purely cosmetic I'm find it better to revert them.

---------

Co-authored-by: Mattias Reichel <mattias.reichel@gmail.com>