Add `grails-plugin-events` to `grails-dependencies` (#13723)
* fix: Add `grails-plugin-events` to `grails-dependencies`

This commit adds the Events plugin to all Grails applications by adding it to the `grails-dependencies`. Previously this has happened transitively through the inclusion of the Async plugin in `grails-dependencies`.

However, as `grails-plugin-async` does not depend on the `grails-plugin-events`, it makes more sense to add `grails-plugin-events` in `grails-dependencies` explicitly.

This prepares for the removal of `grails-plugins-events` to be declared as a dependency in `grails-plugin-async`.

Related:
https://github.com/grails/grails-async/pull/44
https://github.com/grails/grails-async/pull/69
https://github.com/grails/grails-async/pull/70
https://github.com/grails/grails-core/issues/13324

* fix: Remove unnecessary exclusions of `javax:javaee-web-api`

These exclusions should not be necessary anymore as `javax:javaee-web-api` is not included in `grails-plugin-async` or `grails-gsp`.